Some Forge Optimizations and Bug Fixes - Fixed a NullPointer bug that arose when closing the forge without any items in slots. - Cleaned and streamlined ForgeEventListener (GodEventListener). - Removed ForgeManager. - Renamed files and folders to more fitting ones.
