+++ src/main/java/io/github/math0898/rpgframework/main.java
@@ -5,7 +5,7 @@ import io.github.math0898.rpgframework.items.GiveCommand;
 import io.github.math0898.rpgframework.items.ItemManager;
 import io.github.math0898.rpgframework.parties.PartyCommand;
 import io.github.math0898.rpgframework.parties.PartyManager;
-import io.github.math0898.rpgframework.systems.GodEventListener;
+import io.github.math0898.rpgframework.forge.ForgeEventListener;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.event.Listener;
@@ -72,7 +72,7 @@ public final class main extends JavaPlugin implements Listener {
 
         //Register damage listeners
         Bukkit.getPluginManager().registerEvents(new AdvancedDamageHandler(), this);
-        Bukkit.getPluginManager().registerEvents(new GodEventListener(), this); // todo remove me!
+        Bukkit.getPluginManager().registerEvents(new ForgeEventListener(), this); // todo remove me!
         PartyManager.init();
         PlayerManager.init();
         Objects.requireNonNull(Bukkit.getPluginCommand("party")).setExecutor(new PartyCommand());
